+++ admin/app/clusters/[clusterId]/services/ServiceSummary.tsx
@@ -1,30 +1,42 @@
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import Link from "next/link";
+import { SmallLiveGreenCircle } from "../helpers";
 
 const Service = (props: {
   clusterId: string;
   name: string;
   functions?: Array<{ name: string }>;
+  activeFunctions?: Array<{ name: string }>;
 }) => {
   return (
     <Link
       href={`/clusters/${props.clusterId}/services/${props.name}`}
-      className="mr-4 mb-4 w-96"
+      className="mr-2 mb-4 w-96"
     >
-      <Card className="w-[350px] mr-4 mb-4 h-full shadow-sm bg-slate-900 duration-200 border-t-slate-600">
+      <Card className="h-full shadow-sm hover:bg-slate-900 duration-200 border border-green-700 border-t-green-500">
         <CardHeader>
           <p className="text-sm text-gray-500 -mb-1">service</p>
           <CardTitle>{props.name}</CardTitle>
         </CardHeader>
         <CardContent>
-          <p className="text-sm text-gray-500">functions</p>
-          <ul>
-            {props.functions?.map((fn) => (
-              <li className="font-mono" key={fn.name}>
-                {fn.name}()
-              </li>
-            ))}
-          </ul>
+          <p className="text-sm text-gray-500 mb-2">functions</p>
+          <div className="flex flex-col">
+            {props.functions
+              ?.sort((a, b) => (a.name < b.name ? -1 : 1))
+              .map((fn) => (
+                <div
+                  className="flex font-mono py-1 px-2 bg-slate-900 border-slate-700 border mr-2 mb-2 rounded-md items-center justify-between"
+                  key={fn.name}
+                >
+                  <span>{fn.name}()</span>
+                  {props.activeFunctions?.find((f) => f.name === fn.name) && (
+                    <span>
+                      <SmallLiveGreenCircle />
+                    </span>
+                  )}
+                </div>
+              ))}
+          </div>
         </CardContent>
       </Card>
     </Link>
@@ -34,17 +46,20 @@ const Service = (props: {
 export const ServiceSummary = (props: {
   clusterId: string;
   services: Array<{ name: string; functions?: Array<{ name: string }> }>;
+  activeFunctions?: Array<{ name: string; service: string }>;
 }) => {
   return (
     <div className="flex flex-wrap">
-      {props.services.map((s) => (
-        <Service
-          key={s.name}
-          name={s.name}
-          functions={s.functions}
-          clusterId={props.clusterId}
-        />
-      ))}
+      {props.services
+        .sort((a, b) => (a.name < b.name ? -1 : 1))
+        .map((s) => (
+          <Service
+            key={s.name}
+            name={s.name}
+            functions={s.functions}
+            clusterId={props.clusterId}
+          />
+        ))}
     </div>
   );
 };